Refactored styles for login screen
